+++ exthost/action_stress_cpu.go
@@ -6,14 +6,16 @@ package exthost
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"github.com/pkg/errors"
+	"github.com/rs/zerolog/log"
 	"github.com/steadybit/action-kit/go/action_kit_api/v2"
 	"github.com/steadybit/action-kit/go/action_kit_sdk"
 	"github.com/steadybit/extension-host/exthost/resources"
 	"github.com/steadybit/extension-kit/extbuild"
 	"github.com/steadybit/extension-kit/extutil"
 	"strconv"
+	"time"
 )
 
 type stressCPUAction struct{}
@@ -28,12 +30,12 @@ func NewStressCPUAction() action_kit_sdk.Action[resources.StressActionState] {
 	return &stressCPUAction{}
 }
 
-func (l *stressCPUAction) NewEmptyState() resources.StressActionState {
+func (a *stressCPUAction) NewEmptyState() resources.StressActionState {
 	return resources.StressActionState{}
 }
 
 // Describe returns the action description for the platform with all required information.
-func (l *stressCPUAction) Describe() action_kit_api.ActionDescription {
+func (a *stressCPUAction) Describe() action_kit_api.ActionDescription {
 	return action_kit_api.ActionDescription{
 		Id:          fmt.Sprintf("%s.stress-cpu", BaseActionID),
 		Label:       "Stress CPU",
@@ -104,32 +106,15 @@ func (l *stressCPUAction) Describe() action_kit_api.ActionDescription {
 // It must not cause any harmful effects.
 // The passed in state is included in the subsequent calls to start/status/stop.
 // So the state should contain all information needed to execute the action and even more important: to be able to stop it.
-func (l *stressCPUAction) Prepare(_ context.Context, state *resources.StressActionState, request action_kit_api.PrepareActionRequestBody) (*action_kit_api.PrepareResult, error) {
+func (a *stressCPUAction) Prepare(_ context.Context, state *resources.StressActionState, request action_kit_api.PrepareActionRequestBody) (*action_kit_api.PrepareResult, error) {
 	_, err := CheckTargetHostname(request.Target.Attributes)
 	if err != nil {
 		return nil, err
 	}
-	durationConfig := extutil.ToUInt64(request.Config["duration"])
-	if durationConfig < 1000 {
-		return &action_kit_api.PrepareResult{
-			Error: extutil.Ptr(action_kit_api.ActionKitError{
-				Title:  "Duration must be greater / equal than 1s",
-				Status: extutil.Ptr(action_kit_api.Errored),
-			}),
-		}, nil
-	}
-	duration := durationConfig / 1000
-	cpuLoad := extutil.ToUInt(request.Config["cpuLoad"])
-	workers := extutil.ToUInt(request.Config["workers"])
 
-	if cpuLoad == 0 {
-		return nil, errors.New("cpuLoad must be greater than 0")
-	}
-
-	state.StressNGArgs = []string{
-		"--cpu", strconv.Itoa(int(workers)),
-		"--cpu-load", strconv.Itoa(int(cpuLoad)),
-		"--timeout", strconv.Itoa(int(duration)),
+	state.StressNGArgs, err = a.toArgs(request.Config)
+	if err != nil {
+		return nil, err
 	}
 
 	if !resources.IsStressNgInstalled() {
@@ -147,7 +132,7 @@ func (l *stressCPUAction) Prepare(_ context.Context, state *resources.StressActi
 // Start is called to start the action
 // You can mutate the state here.
 // You can use the result to return messages/errors/metrics or artifacts
-func (l *stressCPUAction) Start(_ context.Context, state *resources.StressActionState) (*action_kit_api.StartResult, error) {
+func (a *stressCPUAction) Start(_ context.Context, state *resources.StressActionState) (*action_kit_api.StartResult, error) {
 	return resources.Start(state)
 }
 
@@ -155,6 +140,29 @@ func (l *stressCPUAction) Start(_ context.Context, state *resources.StressAction
 // It will be called even if the start method did not complete successfully.
 // It should be implemented in a immutable way, as the agent might to retries if the stop method timeouts.
 // You can use the result to return messages/errors/metrics or artifacts
-func (l *stressCPUAction) Stop(_ context.Context, state *resources.StressActionState) (*action_kit_api.StopResult, error) {
+func (a *stressCPUAction) Stop(_ context.Context, state *resources.StressActionState) (*action_kit_api.StopResult, error) {
 	return resources.Stop(state)
+}
+
+func (a *stressCPUAction) toArgs(config map[string]interface{}) ([]string, error) {
+	timeout := time.Duration(extutil.ToInt64(config["duration"])) * time.Millisecond
+
+	if timeout < 1*time.Second {
+		return nil, errors.New("Duration must be greater / equal than 1s")
+	}
+
+	workers := extutil.ToInt(config["workers"])
+	cpuLoad := extutil.ToInt(config["cpuLoad"])
+
+	args := []string{
+		"--cpu", strconv.Itoa(workers),
+		"--cpu-load", strconv.Itoa(cpuLoad),
+		"--timeout", strconv.Itoa(int(timeout.Seconds())),
+	}
+
+	if log.Trace().Enabled() {
+		args = append(args, "-v")
+	}
+
+	return args, nil
 }